Move assert event helper to lib
