+++ packages/lib/src/test/index.js
@@ -1,10 +1,12 @@
 import assertions from './helpers/assertions'
+import assertEvent from './helpers/assertEvent'
 import assertRevert from './helpers/assertRevert'
 import shouldBehaveLikeOwnable from './behaviors/Ownable'
 
 const helpers = {
   assertions,
   assertRevert,
+  assertEvent,
 }
 
 const behaviors = {
